kawa: move `Container` and the various layouts to their own file
